requirements.txt
appdirs==1.4.0
elasticsearch==5.2.0
elasticsearch-dsl==5.1.0
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
urllib3==1.20
